re-orginze API trait implementation code
