+++ common/rust/cctrusted_base/src/api.rs
@@ -2,12 +2,7 @@ use crate::api_data::Algorithm;
 use crate::api_data::*;
 use crate::eventlog::TcgEventLog;
 use crate::tcg::TcgDigest;
-use crate::tdx::quote::TdxQuote;
-use crate::tpm::quote::TpmQuote;
-use anyhow::*;
-use core::mem;
 use core::result::Result;
-use core::result::Result::Ok;
 
 pub trait CCTrustedApi {
     /***
@@ -116,24 +111,4 @@ pub trait CCTrustedApi {
 */
 pub trait ParseCcReport<T> {
     fn parse_cc_report(_report: Vec<u8>) -> Result<T, anyhow::Error>;
-}
-
-// API function parses raw cc report to TdxQuote struct
-impl ParseCcReport<TdxQuote> for CcReport {
-    fn parse_cc_report(report: Vec<u8>) -> Result<TdxQuote, anyhow::Error> {
-        match TdxQuote::parse_tdx_quote(report) {
-            Ok(tdx_quote) => unsafe {
-                let report: &TdxQuote = mem::transmute(&tdx_quote);
-                Ok(report.clone())
-            },
-            Err(e) => Err(anyhow!("[parse_cc_report] error parse tdx quote: {:?}", e)),
-        }
-    }
-}
-
-// API function parses raw cc report to TpmQuote struct
-impl ParseCcReport<TpmQuote> for CcReport {
-    fn parse_cc_report(_report: Vec<u8>) -> Result<TpmQuote, anyhow::Error> {
-        todo!()
-    }
 }
